Implemented Operation Document and its children Ration, RegionalRequest, Requisition, Dispatch, Delivery

+++ src/main/java/org/cats/analytics/operation/repository/OperationRepository.java
@@ -1,8 +1,9 @@
 package org.cats.analytics.operation.repository;
 
+import org.bson.types.ObjectId;
 import org.cats.analytics.operation.domain.Operation;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface OperationRepository  extends MongoRepository<Operation, Long> {
-    Operation findByOperationId(int operationId);
+    Operation findByOperationId(Long operationId);
 }
